Convert source to Maven project
