+++ .gitignore
@@ -1,0 +1,5 @@
+/bin/
+/target/
+.classpath
+.project
+/.settings/
